move the countDays function inside the ReleaseDetails page
