+++ packages/core/content-releases/admin/src/pages/ReleaseDetailsPage.tsx
@@ -40,7 +40,6 @@ import {
   useGetReleaseQuery,
   useGetReleaseActionsQuery,
 } from '../services/release';
-import { countDays } from '../utils/countDays';
 /* -------------------------------------------------------------------------------------------------
  * ReleaseDetailsLayout
  * -----------------------------------------------------------------------------------------------*/
@@ -70,6 +69,16 @@ const TrashIcon = styled(Trash)`
     fill: ${({ theme }) => theme.colors.danger600};
   }
 `;
+
+const countDays = (date: string) => {
+  const currentDateInMillis = new Date().getTime();
+  const startDateInMilliseconds = new Date(date).getTime();
+  const timeDiff = currentDateInMillis - startDateInMilliseconds;
+
+  const daysPassed = Math.floor(timeDiff / (1000 * 60 * 60 * 24));
+
+  return daysPassed;
+};
 
 interface ReleaseDetailsLayoutProps {
   toggleEditReleaseModal: () => void;
